# openapi/v2.yaml
openapi: 3.0.3
info:
  title: Santaizi HTTP API
  version: 2.0.0
  description: |
    Santaizi Primary 的公开状态和管理接口。浏览器会话使用 Cookie 与 CSRF，
    自动化客户端使用 Bearer Token。Telemetry/Replication 可靠链路仍使用 gRPC。
servers:
- url: /
  description: 当前站点（同源）
- url: https://monitor.example.com
  description: 示例部署地址
...
security:
- cookieAuth: []
- bearerAuth: []
paths:
...
  /api/v2/admin/telemetry/collectors:
    get:
      tags:
      - Telemetry
      summary: 从端列表
      operationId: listCollectors
      responses:
        '200':
          $ref: '#/components/responses/CollectorListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags:
      - Telemetry
      summary: 创建从端
      operationId: createCollector
      x-ui-route: /admin/telemetry
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        $ref: '#/components/requestBodies/CollectorWrite'
      responses:
        '201':
          $ref: '#/components/responses/CollectorCreatedResponse'
        default:
          $ref: '#/components/responses/Problem'
  /api/v2/admin/telemetry/collectors/{collectorId}:
    parameters:
    - $ref: '#/components/parameters/CollectorId'
    get:
      tags:
      - Telemetry
      summary: 从端详情
      operationId: getCollector
      responses:
        '200':
          $ref: '#/components/responses/CollectorResponse'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      tags:
      - Telemetry
      summary: 更新从端
      operationId: updateCollector
      x-ui-route: /admin/telemetry
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        $ref: '#/components/requestBodies/CollectorWrite'
      responses:
        '200':
          $ref: '#/components/responses/CollectorResponse'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags:
      - Telemetry
      summary: 删除从端
      operationId: deleteCollector
      x-ui-route: /admin/telemetry
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Deleted
        default:
          $ref: '#/components/responses/Problem'
  /api/v2/admin/telemetry/collectors/{collectorId}/rotate-token:
    parameters:
    - $ref: '#/components/parameters/CollectorId'
    post:
      tags:
      - Telemetry
      summary: 轮换从端 Token
      operationId: rotateCollectorToken
      x-ui-route: /admin/telemetry
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          $ref: '#/components/responses/CollectorTokenResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v2/admin/telemetry/collectors/{collectorId}/revoke:
    parameters:
    - $ref: '#/components/parameters/CollectorId'
    post:
      tags:
      - Telemetry
      summary: 撤销从端
      operationId: revokeCollector
      x-ui-route: /admin/telemetry
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          $ref: '#/components/responses/CollectorResponse'
        default:
          $ref: '#/components/responses/Problem'
  /api/v2/admin/telemetry/collectors/{collectorId}/scope:
    parameters:
    - $ref: '#/components/parameters/CollectorId'
    put:
      tags:
      - Telemetry
      summary: 更新从端范围
      operationId: updateCollectorScope
      x-ui-route: /admin/telemetry
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        $ref: '#/components/requestBodies/CollectorScopeWrite'
      responses:
        '200':
          $ref: '#/components/responses/CollectorResponse'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: santaizi-dashboard
      description: 管理员 OAuth 会话 Cookie
    bearerAuth:
      type: http
      scheme: bearer
      description: 'API Token（Authorization: Bearer）'
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 浏览器 Cookie 会话写操作必填；值来自 GET /api/v2/auth/session 的 csrf_token
  parameters:
...
    CollectorId:
      name: collectorId
      in: path
      required: true
      schema:
        type: string
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Cookie 会话写操作时必填；Bearer Token 调用可省略
      schema:
        type: string
  schemas:
    Problem:
      type: object
      required:
      - title
      - status
      - code
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        trace_id:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Meta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
          format: int64
        next_cursor:
          type: string
...
    CollectorToken:
      type: object
      required:
      - collector_id
      - registration_token
      - revoked
      properties:
        collector_id:
          type: string
        registration_token:
          type: string
        revoked:
          type: boolean
...
    CollectorCreated:
      type: object
      required:
      - collector
      - registration_token
      properties:
        collector:
          $ref: '#/components/schemas/Collector'
        registration_token:
          type: string
...
    Collector:
      type: object
      required:
      - id
      - name
      - address
      - tls
      - insecure_tls
      - generation
      - config_version
      - revoked
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
          description: 探针访问地址（host:访问端口）
        listen_port:
          type: integer
          minimum: 0
          maximum: 65535
          description: 从端进程监听端口。0 表示与 address 中的访问端口相同。
        tls:
          type: boolean
        insecure_tls:
          type: boolean
        location:
          type: string
          maxLength: 64
          description: 从端在地球上的位置，ISO 国家/地区码或 `lat,lon`
        kind:
          type: string
          enum:
          - observer
          - probe
          description: 从端类型。创建后不可改。缺省 observer。
        probe_interval_seconds:
          type: integer
          minimum: 1
        mtr_interval_seconds:
          type: integer
          minimum: 1
        mtr_probes:
          type: integer
          minimum: 1
          maximum: 10
        tcp_ports:
          type: string
          description: 逗号分隔 TCP 端口，如 22,443
        enable_icmp:
          type: boolean
        enable_tcp:
          type: boolean
        enable_mtr:
          type: boolean
        enable_ipv4:
          type: boolean
          default: true
          description: 探测 IPv4。与 enable_ipv6 至少开一项。
        enable_ipv6:
          type: boolean
          default: true
          description: 探测 IPv6。与 enable_ipv4 至少开一项。
        route_interval_seconds:
          type: integer
          enum:
          - 3600
          - 86400
          - 604800
        route_keep:
          type: integer
          minimum: 1
          maximum: 50
        notify:
          type: boolean
        notification_tag:
          type: string
        latency_notify:
          type: boolean
        min_latency_ms:
          type: number
          format: double
        max_latency_ms:
          type: number
          format: double
        fail_threshold:
          type: integer
          minimum: 1
        generation:
          type: integer
          format: int64
        config_version:
          type: integer
          format: int64
        revoked:
          type: boolean
        status:
          type: string
          enum:
          - online
          - offline
          - unknown
        last_seen:
          type: string
          format: date-time
        last_sync:
          type: string
          format: date-time
        last_primary_seen:
          type: string
          format: date-time
        spool_size:
          type: integer
          format: int64
        pending_records:
          type: integer
          format: int64
        oldest_pending:
          type: string
          format: date-time
        replication_cursor:
          type: integer
          format: int64
        connected_agents:
          type: integer
          format: int64
        protocol_version:
          type: string
        software_version:
          type: string
          description: 从端二进制版本
        heartbeat_rtt_ms:
          type: number
          format: double
        heartbeat_rtt_sampled_at:
          type: string
          format: date-time
        replication_rtt_ms:
          type: number
          format: double
        replication_rtt_sampled_at:
          type: string
          format: date-time
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/CollectorScope'
    CollectorScope:
      type: object
      required:
      - type
      - value
      properties:
        type:
          type: string
          enum:
          - all
          - server
          - group
          - tag
        value:
          type: string
...
    CollectorWrite:
      type: object
      required:
      - name
      - scopes
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        kind:
          type: string
          enum:
          - observer
          - probe
          description: 仅创建时生效；更新忽略。
        address:
          type: string
          maxLength: 255
          description: 探针访问地址（host:访问端口）。探测型可空。
        listen_port:
          type: integer
          minimum: 0
          maximum: 65535
          description: 从端进程监听端口。0 表示与 address 中的访问端口相同。
        tls:
          type: boolean
        insecure_tls:
          type: boolean
        location:
          type: string
          maxLength: 64
        probe_interval_seconds:
          type: integer
          minimum: 1
        mtr_interval_seconds:
          type: integer
          minimum: 1
        mtr_probes:
          type: integer
          minimum: 1
          maximum: 10
        tcp_ports:
          type: string
        enable_icmp:
          type: boolean
        enable_tcp:
          type: boolean
        enable_mtr:
          type: boolean
        enable_ipv4:
          type: boolean
          default: true
          description: 探测 IPv4。与 enable_ipv6 至少开一项。
        enable_ipv6:
          type: boolean
          default: true
          description: 探测 IPv6。与 enable_ipv4 至少开一项。
        route_interval_seconds:
          type: integer
          enum:
          - 3600
          - 86400
          - 604800
        route_keep:
          type: integer
          minimum: 1
          maximum: 50
        notify:
          type: boolean
        notification_tag:
          type: string
        latency_notify:
          type: boolean
        min_latency_ms:
          type: number
          format: double
        max_latency_ms:
          type: number
          format: double
        fail_threshold:
          type: integer
          minimum: 1
        scopes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CollectorScope'
    CollectorScopeWrite:
      type: object
      required:
      - scopes
      properties:
        scopes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CollectorScope'
...
  requestBodies:
...
    CollectorWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CollectorWrite'
    CollectorScopeWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CollectorScopeWrite'
...
  responses:
    Problem:
      description: Problem details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    CollectorResponse:
      description: Collector
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            properties:
              data:
                $ref: '#/components/schemas/Collector'
    CollectorCreatedResponse:
      description: Created collector and its registration token
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            properties:
              data:
                $ref: '#/components/schemas/CollectorCreated'
    CollectorListResponse:
      description: Collector list
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            - meta
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/Collector'
              meta:
                $ref: '#/components/schemas/Meta'
...
    CollectorTokenResponse:
      description: Collector registration token
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            properties:
              data:
                $ref: '#/components/schemas/CollectorToken'
...